+++ TrinketTinker/Models/MotionArgs/HoverArgs.cs
@@ -9,7 +9,7 @@ public sealed class HoverArgs : LerpArgs
     public float Magnitude { get; set; } = 16f;
 
     /// <summary>Period of bobbing motion, in ms</summary>
-    public float Period { get; set; } = 800f;
+    public double Period { get; set; } = 800f;
 
     /// <summary>If set, the companion will perch on the player's head after the player stays still for this many miliseconds</summary>
     public float? PerchingTimeout { get; set; } = null;
